+++ lib/src/subtypes/races/orc/orc.dart
@@ -14,14 +14,16 @@ import '../../../generators/npcs/physical/hair/base_hair_generator.dart';
 import '../../../generators/npcs/physical/special_features/base_special_feature_generator.dart';
 import '../../../generators/world/opinions/base_opinion_generator.dart';
 import '../race.dart';
+import '../settlement_nameable.dart';
 import 'alignment/orc_alignment_generator.dart';
 import 'names/orc_name_generator.dart';
 import 'opinions/orc_opinion_generator.dart';
 import 'quirks/orc_quirks_generator.dart';
+import 'settlement_names/orcish_settlement_name_generator.dart';
 import 'skin/orc_skin_generator.dart';
 
 /// A class that represents the orc race
-class Orc implements Race {
+class Orc implements Race, SettlementNameable {
   const Orc();
 
   static const _raceName = "orc";
@@ -89,4 +91,8 @@ class Orc implements Race {
         BaseOpinionGenerator(),
         OrcOpinionGenerator(),
       ]);
+
+  @override
+  IGenerator<String> getSettlementNameGenerator() =>
+      OrcishSettlementNameGenerator();
 }